Add progress and color extensions in general extensionMethods file
