--- SourceCode/GPS/Classes/CColorExtensions.cs
@@ -1,20 +1,0 @@
-﻿using System.Drawing;
-
-namespace AgOpenGPS
-{
-    public static class CColorExtensions
-    {
-        public static Color CheckColorFor255(this Color color)
-        {
-            var currentR = color.R;
-            var currentG = color.G;
-            var currentB = color.B;
-
-            if (currentR == 255) currentR = 254;
-            if (currentG == 255) currentG = 254;
-            if (currentB == 255) currentB = 254;
-
-            return Color.FromArgb(color.A, currentR, currentG, currentB);
-        }
-    }
-}
