+++ SourceCode/GPS/Classes/CExtensionMethods.cs
@@ -1,0 +1,45 @@
+﻿using System.Drawing;
+using System.Windows.Forms;
+
+namespace AgOpenGPS
+{
+    public static class CExtensionMethods
+    {
+        /// <summary>
+        /// Sets the progress bar value, without using 'Windows Aero' animation.
+        /// This is to work around a known WinForms issue where the progress bar 
+        /// is slow to update. 
+        /// </summary>
+        public static void SetProgressNoAnimation(this ProgressBar pb, int value)
+        {
+            // To get around the progressive animation, we need to move the 
+            // progress bar backwards.
+            if (value == pb.Maximum)
+            {
+                // Special case as value can't be set greater than Maximum.
+                pb.Maximum = value + 1;     // Temporarily Increase Maximum
+                pb.Value = value + 1;       // Move past
+                pb.Maximum = value;         // Reset maximum
+            }
+            else
+            {
+                pb.Value = value + 1;       // Move past
+            }
+            pb.Value = value;               // Move to correct value
+        }
+
+        public static Color CheckColorFor255(this Color color)
+        {
+            var currentR = color.R;
+            var currentG = color.G;
+            var currentB = color.B;
+
+            if (currentR == 255) currentR = 254;
+            if (currentG == 255) currentG = 254;
+            if (currentB == 255) currentB = 254;
+
+            return Color.FromArgb(color.A, currentR, currentG, currentB);
+        }
+    }
+
+}
